Refactor /account into subfiles
